requirements.txt
boto3==1.16.40
Passlib==1.7.1
itsdangerous==2.0.1
Flask==2.2.0
Flask-Caching==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.3
Flask-Migrate==3.1.0
flask-marshmallow==0.14.0
Flask_Cors==3.0.9
Redis==3.5.3
requests==2.25.1
SQLAlchemy==1.4.42
waitress==1.4.4
marshmallow-sqlalchemy==0.24.1
PyMySQL==0.10.1
python-dotenv==0.15.0
python-slugify
cryptography
xlrd==1.2.0
pytz
debugpy